Add C style interface to the API for pixel binning and tests in the API test code.

+++ include/api/seabreezeapi/SpectrometerFeatureAdapter.h
@@ -65,7 +65,7 @@ class SpectrometerFeatureAdapter: public FeatureAdapterTemplate<OOISpectrometerF
 		unsigned long integrationTimeMicros);
 	long getMinimumIntegrationTimeMicros(int *errorCode);
 	long getMaximumIntegrationTimeMicros(int *errorCode);
-	int getMaximumIntensity(int *errorCode);
+	double getMaximumIntensity(int *errorCode);
 };
 
 }// namespace api
